:truck: chore: wrapped up all ui in shared core
