+++ app/src/main/java/com/shub39/grit/app/MainAppState.kt
@@ -14,7 +14,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
-package com.shub39.grit.core.domain
+package com.shub39.grit.app
 
 import androidx.compose.runtime.Immutable
 import androidx.compose.runtime.Stable
