+++ app/src/main/java/com/shub39/grit/viewmodels/MainViewModel.kt
@@ -20,10 +20,10 @@ import androidx.compose.ui.graphics.Color
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.shub39.grit.BuildConfig
+import com.shub39.grit.app.MainAppState
 import com.shub39.grit.billing.BillingHandler
 import com.shub39.grit.billing.SubscriptionResult
 import com.shub39.grit.core.data.ChangelogManager
-import com.shub39.grit.core.domain.MainAppState
 import com.shub39.grit.core.domain.SettingsDatastore
 import com.shub39.grit.core.domain.ThemeDatastore
 import kotlinx.coroutines.Job
