+++ shared/core/src/commonMain/kotlin/com/shub39/grit/core/app/AppSections.kt
@@ -1,0 +1,57 @@
+/*
+ * Copyright (C) 2026  Shubham Gorai
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+package com.shub39.grit.core.app
+
+import androidx.navigation3.runtime.NavKey
+import grit.shared.core.generated.resources.Res
+import grit.shared.core.generated.resources.alarm
+import grit.shared.core.generated.resources.check_list
+import grit.shared.core.generated.resources.habits
+import grit.shared.core.generated.resources.settings
+import grit.shared.core.generated.resources.tasks
+import kotlinx.serialization.Serializable
+import org.jetbrains.compose.resources.DrawableResource
+import org.jetbrains.compose.resources.StringResource
+
+@Serializable
+sealed interface AppSections : NavKey {
+    @Serializable data object HabitsPages : AppSections
+
+    @Serializable data object TaskPages : AppSections
+
+    @Serializable data object SettingsPages : AppSections
+
+    companion object {
+        val mainRoutes = listOf(TaskPages, HabitsPages, SettingsPages)
+
+        fun AppSections.toStringRes(): StringResource {
+            return when (this) {
+                HabitsPages -> Res.string.habits
+                TaskPages -> Res.string.tasks
+                SettingsPages -> Res.string.settings
+            }
+        }
+
+        fun AppSections.toIconRes(): DrawableResource {
+            return when (this) {
+                HabitsPages -> Res.drawable.alarm
+                TaskPages -> Res.drawable.check_list
+                SettingsPages -> Res.drawable.settings
+            }
+        }
+    }
+}
